notification_service/requirements.txt
kafka-python-ng==2.2.3
prometheus_client==0.21.0
python-dotenv==1.0.1
